[pipeline] Refactor the pipeline to split out reusable code
Move the repo independent parts (ci-prepare, ci-start, ci-cleanup) to a
template and leave in the main job only the actual tests (script).

The created template is meant to be moved to mendertesting repo and be
reused by other repos.

Minor functional tweaks to the original scripts, the design is kept as
it was.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>