[pipeline] Take the ci-cleanup script out of the template
The technical detail here is that the way we start Docker in the
before_script (running dockerd in the background) makes the CI kill the
process on exit, and hence the clean-up will fail in after_script trying
to contact Docker.

The clean-up is not strictly skeaking necessary in the CI, so just
remove the funcionality.

Leave the script in the repo for future reference or for manual runs.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>