[ci-make-clean] Remove everything from current namespace
In addition to uninstalling the helm charts, remove also any other
resource that could have been crated via kubectl (like nats and the
tester pod).

Also, let the script continue on failure so that it gets to the new call
even if helm uninstall fails (for example if no charts are installed).

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>