Pass non-SSL env variable to api-gateway
If SSL termination is handled outdide of the api-gateway pod, this
state shuold be signaled using the `SSL` flag. Since the api-gateway
container already implements this infrastructure requirement and
handles the SSL environment variable, it is just a matter of
passing the SSL variable from the helm values to the container.

Signed-off-by: Alexandru Ionita <alexandru.ionita@gmail.com>