values: optionally set root CA cert
self signed certs can still be used, but they must
not be 0-depth anymore. they must have a signing authority,
even if it's a bogus CA.

if we're using a CA - it must be added to deployments' cert store.
there is an env var for it in the service, but we need settings and
a secret in helm.

this is a side effect of either migrating to openssh (client), or
bumping go (client, deployments), not sure which.

Issues: MEN-4679

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>