minio routing
this is to enable using minio at all:

- conditionally add traefik routing for minio
- add supporting config to default values

note we're using just a single domain here - traefik will
handle storage as well. this removes the need to upfront
set up 2 domains, for which we don't really have an explanation
(leftover from initial setups with storage proxy).

we need some discrimination mechanism for routing, so it's a port.

the original PathPrefix(/bucket) is not enough, deployments pings this
bucket only on startup. on health checks, it pings the `/` route (get
buckets).

the original Host || X-Forwarded-Host rule is also not enough, as it would
effectively route the health check to the ui.

conclusion: use separate port, single domain for any storage path.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>