feat: optionals deviceauth cronjobs
You can create custom cronjobs based on the deviceauth service.
This feature is going to deprecate the device_license_count feature,
that could be installed as a custom cronjob, with the provided example
in the Values file.
The device_license_count key will be removed in the future.

Ticket: MC-7789

Signed-off-by: Roberto Giovanardi <roberto.giovanardi@northern.tech>