feat: included dependency from bitnami to bitnamisecure
By default this helm chart comes with convinient out-of-the-box
dependencies from Bitnami: Redis and MongoDB. These dependencies are not
supposed to be used in production.
The dependencies was moved from the legacy bitnami repository to the new
bitnamisecure repository with only the latest tag available.

Please don't use these dependencies in production, but for testing
purposes only.

Ticket: MC-8173

Signed-off-by: Roberto Giovanardi <roberto.giovanardi@northern.tech>